Input protocol v2: implement native-side shared byte parser
vs_host_bridge.c/rg_input.c were still reading the wire one raw byte at
a time and using it directly as host_buttons -- the MicroPython and
emulator sides moved to the v2 framing (docs/input-protocol-v2.md) in
aa36f4b, but the native retro-go apps (prboom-go, retro-core) never
got the matching update, so their input was never functional against
either the base station or the workbench.

Add a C mirror of input_parser.py's state machine (VS_SCAN/VS_JOY/VS_CMD)
in vs_host_bridge.c, draining all buffered bytes per poll instead of one
byte at a time, and handle "reset"/"ota_start <url>" command frames
in-band. rg_input.c now mirrors extra bit 0 into host_buttons bit 7 so
BUTTON_D/RG_KEY_START keeps working now that it travels off-wire.

+++ components/retro-go/rg_input.h
@@ -74,7 +74,7 @@ typedef struct
 typedef struct
 {
     rg_key_t key;
-    uint8_t mask; // bit mask in the most recently received host byte
+    uint32_t mask; // bit mask in the most recently received host state (joy1 bits 0-7; joy2/extra reserved for bits 8+)
 } rg_keymap_host_t;
 
 // #define RG_GAMEPAD_VIRT_MAP {{}, ...} to add virtual buttons (eg start+select = menu)
